Extracted a few more functions out of the cache refresher
